1) Some pieces of the code related to experience have been rewritten and improved. Amongst other things, now the number of experience points given by a creature is treated as a double and rounded as an integer only at the last step of adding those experience points to a player. It gives a better result, because it avoids the issue of losing too much information while rounding numbers (which was causing things like a rat giving only 200+200 of xp in a x100 server); 2) From now on, you can control how much of the experience won by a summon is given to its master player through the variable summons_give_exp at config.lua; 3) eliminated the combat::round method (it is better to use the round function from math.h); 4) Corrected the id (and decay) of the cheese and the scarab cheese at source code and lua files (including loot monsters); 5) Player::getStaminaMinutes is now a const method, as it should be.

+++ src/combat.cpp
@@ -1449,8 +1449,8 @@ void AreaCombat::copyArea(const MatrixArea* input, MatrixArea* output, MatrixOpe
 				int32_t newY = y - centerY;
 
 				//perform rotation
-				int32_t rotatedX = round(newX * a + newY * b);
-				int32_t rotatedY = round(newX * c + newY * d);
+				int32_t rotatedX = (int32_t) round(newX * a + newY * b);
+				int32_t rotatedY = (int32_t) round(newX * c + newY * d);
 
 				//write in the output matrix using rotated coordinates
 				(*output)[rotatedY + rotateCenterY][rotatedX + rotateCenterX] = (*input)[y][x];
